Refactored code in train.py and cleaned up utils.py
