+++ experiments/dmft/utils.py
@@ -1,105 +1,23 @@
+"""
+Shared training utilities live in ``experiments.limits_paper.utils``. This
+module overrides the BP ``MLP`` so µPC weights are N(0, 1), matching
+``jpc.make_mlp``, and hosts the finite-size PC/BP training loops.
+"""
+
 import os
-import numpy as np
+import shutil
+from pathlib import Path
+
+import equinox as eqx
+import jax
 import jax.numpy as jnp
 import jax.random as jr
-from jax.flatten_util import ravel_pytree
-from jax import vmap
+import numpy as np
+import optax
 
 import jpc
-import optax
-import equinox as eqx
-import equinox.nn as nn
-
-
-def configure_param_optim(optim_id, param_type, use_skips, param_lr, width, depth, gamma_0=1.0):
-    if param_type == "sp":
-        return optax.sgd(param_lr) if optim_id == "gd" else optax.adam(param_lr)
-    else:
-        if optim_id == "gd":
-            scaled_lr = param_lr * (gamma_0**2) * width
-            return optax.sgd(scaled_lr)
-        elif optim_id == "adam":
-            if use_skips:
-                scaled_lr = param_lr / ( np.sqrt(width) * np.sqrt(depth) )
-            else:
-                scaled_lr = param_lr / np.sqrt(width)
-            return optax.adam(scaled_lr)
-        else:
-            raise ValueError(f"Invalid optimiser: {optim_id}")
-    
-
-def setup_pc_experiment(
-        results_dir,
-        input_dim,
-        n_samples,
-        n_hidden,
-        use_skips,
-        act_fn,
-        param_type,
-        param_optim_id,
-        param_lr,
-        gamma_0,
-        n_train_iters,
-        infer_mode,
-        n_infer_iters,
-        activity_lr,
-        width,
-        loss_id,
-        seed
-):
-    return os.path.join(
-        results_dir,
-        f"{input_dim}_input_dim",
-        f"{n_samples}_n_samples",
-        f"{n_hidden}_n_hidden",
-        f"{use_skips}_use_skips",
-        f"{act_fn}_act_fn",
-        f"{param_type}_param_type",
-        f"{param_optim_id}_param_optim",
-        f"{param_lr}_param_lr",
-        f"{gamma_0}_gamma_0",
-        f"{n_train_iters}_n_train_iters",
-        f"{infer_mode}_infer_mode",
-        f"{n_infer_iters}_n_infer_iters",
-        f"{activity_lr}_activity_lr",
-        f"{width}_width",
-        f"{loss_id}_loss_id",
-        str(seed)
-    )
-
-
-def setup_bp_experiment(
-        results_dir,
-        input_dim,
-        n_samples,
-        n_hidden,
-        use_skips,
-        act_fn,
-        param_type,
-        optim_id,
-        param_lr,
-        gamma_0,
-        n_train_iters,
-        width,
-        loss_id,
-        seed
-):
-    return os.path.join(
-        results_dir,
-        f"{input_dim}_input_dim",
-        f"{n_samples}_n_samples",
-        f"{n_hidden}_n_hidden",
-        f"{use_skips}_use_skips",
-        f"{act_fn}_act_fn",
-        f"{param_type}_param_type",
-        f"{optim_id}_optim_id",
-        f"{param_lr}_param_lr",
-        f"{gamma_0}_gamma_0",
-        f"{n_train_iters}_n_train_iters",
-        f"{width}_width",
-        f"{loss_id}_loss_id",
-        str(seed)
-    )
+from experiments.limits_paper.utils import MLP as LimitsMLP
+from experiments.limits_paper.utils import configure_param_optim, flatten_grads
 
 
 def create_toy_dataset(key, D, P):
@@ -108,15 +26,7 @@ def create_toy_dataset(key, D, P):
     return X, y
 
 
-class MLP(eqx.Module):
-    D: int
-    N: int
-    L: int
-    param_type: str
-    use_skips: bool
-    layers: list
-    gamma: float
-
+class MLP(LimitsMLP):
     def __init__(
             self,
             key,
@@ -130,146 +40,273 @@ class MLP(eqx.Module):
             use_bias=False,
             use_skips=False
     ):
-        self.D = d_in
-        self.N = N
-        self.L = L
-        self.param_type = param_type
-        self.use_skips = use_skips
-        self.gamma = gamma
-
-        keys = jr.split(key, L)
-        self.layers = []
-        for i in range(L):
-            act_fn_l = nn.Identity() if i == 0 else jpc.get_act_fn(act_fn)
-            _in = d_in if i == 0 else N
-            _out = d_out if (i + 1) == L else N
-            linear = nn.Linear(
-                _in,
-                _out,
-                use_bias=use_bias,
-                key=keys[i]
-            )
-            # µPC applies explicit forward scalings, so weights must be N(0, 1)
-            # (same convention as jpc.make_mlp). Equinox's default 1/sqrt(fan_in)
-            # init would otherwise double-scale and kill learning.
-            if param_type == "mupc":
-                W = jr.normal(keys[i], linear.weight.shape)
-                linear = eqx.tree_at(lambda l: l.weight, linear, W)
-            layer = nn.Sequential(
-                [
-                    nn.Lambda(act_fn_l),
-                    linear,
-                ]
-            )
-            self.layers.append(layer)
-
-    def __call__(self, x):
-        if self.param_type == "mupc":
-            for i, f in enumerate(self.layers):
-                if (i + 1) == 1:
-                    x = f(x) / jnp.sqrt(self.D)
-                elif 1 < (i + 1) < self.L:
-                    residual = x if self.use_skips else 0
-                    rescaling = jnp.sqrt(
-                        self.N * self.L
-                    ) if self.use_skips else jnp.sqrt(self.N)
-                    x = (f(x) / rescaling) + residual
-                elif (i + 1) == self.L:
-                    x = f(x) / self.N
-
-        else:
-            for i, f in enumerate(self.layers):
-                residual = x if self.use_skips and (1 < (i + 1) < self.L) else 0
-
-                x = f(x) + residual
-
-        return x / self.gamma
-
-
-def flatten_grads(grads):
-        flat_grads, _ = ravel_pytree(eqx.filter(grads, eqx.is_array))
-        return flat_grads
-
-
-def compute_grad_cosine_similarities(pc_grads, bp_grads):
-    """
-    Compute cosine similarities between PC and BP gradients over multiple training steps.
-    Memory-efficient implementation that processes gradients iteratively to avoid OOM.
-    
-    Args:
-        pc_grads: List or array of PC gradients (shape: n_train_iters x grad_dim)
-        bp_grads: List or array of BP gradients (shape: n_train_iters x grad_dim)
-    
-    Returns:
-        Array of cosine similarities (shape: n_train_iters)
-    """
-    # Ensure same number of iterations
-    n_iterations = min(len(pc_grads), len(bp_grads))
-    
-    # Pre-allocate output array
-    cosine_similarities = np.zeros(n_iterations)
-    
-    # Process each iteration separately to avoid loading all gradients into memory
-    for i in range(n_iterations):
-        # Get gradients for this iteration (already numpy arrays from train functions)
-        pc_grad = pc_grads[i]
-        bp_grad = bp_grads[i]
-        
-        # Ensure they are 1D arrays
-        if pc_grad.ndim > 1:
-            pc_grad = pc_grad.flatten()
-        if bp_grad.ndim > 1:
-            bp_grad = bp_grad.flatten()
-        
-        # Ensure same gradient dimension
-        min_dim = min(len(pc_grad), len(bp_grad))
-        pc_grad = pc_grad[:min_dim]
-        bp_grad = bp_grad[:min_dim]
-        
-        # Compute cosine similarity: cos_sim = (a · b) / (||a|| * ||b||)
-        dot_product = np.dot(pc_grad, bp_grad)
-        pc_norm = np.linalg.norm(pc_grad)
-        bp_norm = np.linalg.norm(bp_grad)
-        
-        # Avoid division by zero
-        norms_product = pc_norm * bp_norm
-        if norms_product > 1e-10:
-            cosine_similarities[i] = dot_product / norms_product
-        else:
-            cosine_similarities[i] = 0.0
-        
-        # Explicitly delete references to free memory (though Python GC should handle this)
-        del pc_grad, bp_grad
-    
-    return cosine_similarities
-
-
-def evaluate_pc(params, test_loader, param_type):
-    model, skip_model = params
-    avg_test_loss, avg_test_acc = 0, 0
-    for _, (img_batch, label_batch) in enumerate(test_loader):
-        img_batch, label_batch = img_batch.numpy(), label_batch.numpy()
-
-        test_loss, test_acc = jpc.test_discriminative_pc(
-            model=model,
-            output=label_batch,
-            input=img_batch,
-            skip_model=skip_model,
-            param_type=param_type
+        super().__init__(
+            key,
+            d_in,
+            N,
+            L,
+            d_out,
+            act_fn,
+            param_type,
+            gamma,
+            use_bias=use_bias,
+            use_skips=use_skips,
         )
-        avg_test_loss += test_loss
-        avg_test_acc += test_acc
+        # µPC applies explicit forward scalings, so weights must be N(0, 1)
+        # (same convention as jpc.make_mlp). Equinox's default 1/sqrt(fan_in)
+        # init would otherwise double-scale and kill learning.
+        if param_type != "mupc":
+            return
+        keys = jr.split(key, L)
+        layers = []
+        for i, layer in enumerate(self.layers):
+            linear = layer[1]
+            W = jr.normal(keys[i], linear.weight.shape)
+            linear = eqx.tree_at(lambda l: l.weight, linear, W)
+            layers.append(eqx.tree_at(lambda s: s[1], layer, linear))
+        object.__setattr__(self, "layers", layers)
 
-    return avg_test_loss / len(test_loader), avg_test_acc / len(test_loader)
+
+def get_output_energy_scaling(param_type: str, gamma_0: float, width: int) -> float:
+    """Match ``test_coord_check.get_coord_data`` µPC output-energy scaling."""
+    return (gamma_0 ** 2) * width if param_type == "mupc" else 1.0
 
 
-def evaluate_bp(model, test_loader):
-    avg_test_loss, avg_test_acc = 0, 0
-    for _, (img_batch, label_batch) in enumerate(test_loader):
-        img_batch, label_batch = img_batch.numpy(), label_batch.numpy()
-        
-        y_pred = vmap(model)(img_batch)
-        avg_test_loss += 0.5 * jnp.mean(jnp.sum((label_batch - y_pred) ** 2, axis=1))
-        avg_test_acc += jpc.compute_accuracy(label_batch, y_pred)
-    
-    return avg_test_loss / len(test_loader), avg_test_acc / len(test_loader)
+def cleanup_experiment_dirs(results_dir: str):
+    """Remove finite-sim result trees (``*_input_dim``), keeping plot pngs."""
+    removed = []
+    root = Path(results_dir)
+    if not root.exists():
+        return removed
+    for path in sorted(root.glob("*_input_dim")):
+        if path.is_dir():
+            shutil.rmtree(path)
+            removed.append(str(path))
+    return removed
+
+
+def train_pcn(
+      model,
+      use_skips,
+      X_input,
+      Y_target,
+      width,
+      gamma_0,
+      param_type,
+      infer_mode,
+      n_infer_iters,
+      activity_lr,
+      param_optim_id,
+      param_lr,
+      n_train_iters,
+      loss_id,
+      save_dir,
+      store_grads=False
+):
+    """Train a PC network.
+
+    Parameter / activity updates follow the finite-size convention used by
+    ``get_coord_data``: plain ``param_lr`` with
+    ``output_energy_scaling = gamma^2 * width`` for µPC (rather than baking the
+    width factor into the optimiser learning rate).
+    """
+    os.makedirs(save_dir, exist_ok=True)
+
+    depth = len(model)
+    skip_model = jpc.make_skip_model(depth) if use_skips else None
+    output_energy_scaling = get_output_energy_scaling(param_type, gamma_0, width)
+
+    # Optimisers (plain lr; µPC width/gamma scaling via output_energy_scaling)
+    batch_size = X_input.shape[0]
+    activity_optim = optax.sgd(activity_lr * batch_size)
+    if param_optim_id == "gd":
+        param_optim = optax.sgd(param_lr)
+    elif param_optim_id == "adam":
+        param_optim = optax.adam(param_lr)
+    else:
+        raise ValueError(f"Invalid optimiser: {param_optim_id}")
+    param_opt_state = param_optim.init(
+        (eqx.filter(model, eqx.is_array), skip_model)
+    )
+
+    num_energies, theory_energies = [], []
+    train_losses = []
+    loss_rescalings = []
+    pc_grads = [] if store_grads else None
+
+    for _ in range(n_train_iters):
+
+        # Record supervised loss on the current feedforward prediction *before*
+        # the parameter update, matching get_coord_data / DMFT step indexing.
+        activities = jpc.init_activities_with_ffwd(
+            model=model,
+            input=X_input,
+            skip_model=skip_model,
+            param_type=param_type,
+            gamma=gamma_0
+        )
+        if loss_id == "mse":
+            train_loss = jpc.mse_loss(activities[-1], Y_target)
+        else:
+            train_loss = jpc.cross_entropy_loss(activities[-1], Y_target)
+        train_losses.append(train_loss)
+
+        if infer_mode == "closed_form":
+            equilib_energy, S = jpc.linear_equilib_energy(
+                params=(model, skip_model),
+                x=X_input,
+                y=Y_target,
+                param_type=param_type,
+                gamma=gamma_0,
+                return_rescaling=True,
+                output_energy_scaling=output_energy_scaling,
+            )
+            theory_energies.append(equilib_energy)
+            loss_rescaling = jnp.linalg.norm(S, ord=2) if Y_target.ndim > 1 else S
+            loss_rescalings.append(loss_rescaling)
+
+        # inference
+        if infer_mode == "optim":
+            activity_opt_state = activity_optim.init(activities)
+            for _ in range(n_infer_iters):
+                activity_update_result = jpc.update_pc_activities(
+                    params=(model, skip_model),
+                    activities=activities,
+                    optim=activity_optim,
+                    opt_state=activity_opt_state,
+                    output=Y_target,
+                    input=X_input,
+                    param_type=param_type,
+                    gamma=gamma_0,
+                    loss_id=loss_id,
+                    output_energy_scaling=output_energy_scaling,
+                )
+                activities = activity_update_result["activities"]
+                activity_opt_state = activity_update_result["opt_state"]
+                energy = activity_update_result["energy"]
+
+            num_energies.append(energy)
+
+            param_update_result = jpc.update_pc_params(
+                params=(model, skip_model),
+                activities=activities,
+                optim=param_optim,
+                opt_state=param_opt_state,
+                output=Y_target,
+                input=X_input,
+                param_type=param_type,
+                gamma=gamma_0,
+                loss_id=loss_id,
+                output_energy_scaling=output_energy_scaling,
+            )
+
+        else:
+            # learning with closed form energy
+            param_update_result = jpc.update_linear_equilib_energy_params(
+                params=(model, skip_model),
+                optim=param_optim,
+                opt_state=param_opt_state,
+                y=Y_target,
+                x=X_input,
+                param_type=param_type,
+                gamma=gamma_0,
+                output_energy_scaling=output_energy_scaling,
+            )
+
+        model = param_update_result["model"]
+        skip_model = param_update_result["skip_model"]
+        param_opt_state = param_update_result["opt_state"]
+        grads = param_update_result["grads"]
+
+        if pc_grads is not None:
+            flat_grads = flatten_grads(grads)
+            # Convert JAX array to numpy immediately to free memory
+            pc_grads.append(np.array(flat_grads))
+            del flat_grads, grads
+
+    energies = (
+        jnp.array(theory_energies)
+        if infer_mode == "closed_form"
+        else jnp.array(num_energies)
+    )
+    np.save(f"{save_dir}/energies.npy", energies)
+    np.save(f"{save_dir}/train_losses.npy", np.array(train_losses))
+    np.save(f"{save_dir}/loss_rescalings.npy", loss_rescalings)
+
+    return pc_grads
+
+
+def train_bpn(
+      model,
+      use_skips,
+      X_input,
+      Y_target,
+      width,
+      gamma_0,
+      param_type,
+      optim_id,
+      param_lr,
+      n_train_iters,
+      loss_id,
+      save_dir,
+      store_grads=False
+):
+    os.makedirs(save_dir, exist_ok=True)
+
+    # Optimiser
+    optim = configure_param_optim(
+        optim_id, param_type, use_skips, param_lr, width, model.L, gamma_0
+    )
+    opt_state = optim.init(eqx.filter(model, eqx.is_array))
+
+    if loss_id == "mse":
+        @eqx.filter_jit
+        def loss_fn(model, x, y):
+            y_pred = jax.vmap(model)(x)
+            return 0.5 * jnp.mean(jnp.sum((y - y_pred) ** 2, axis=1))
+    else:
+        @eqx.filter_jit
+        def loss_fn(model, x, y):
+            y_pred = jax.vmap(model)(x)
+            return jpc.cross_entropy_loss(y_pred, y)
+
+    @eqx.filter_jit
+    def make_step(model, optim, opt_state, x, y):
+        loss, grads = eqx.filter_value_and_grad(loss_fn)(model, x, y)
+        updates, opt_state = optim.update(
+            updates=grads,
+            state=opt_state,
+            params=eqx.filter(model, eqx.is_array)
+        )
+        model = eqx.apply_updates(model, updates)
+        return model, opt_state, loss, grads
+
+    losses = []
+    bp_grads = [] if store_grads else None
+
+    for _ in range(n_train_iters):
+        # Record loss before the parameter update to match get_Delta / DMFT
+        # step indexing (pre-update residual).
+        if loss_id == "mse":
+            y_pred = jax.vmap(model)(X_input)
+            train_loss = float(
+                0.5 * jnp.mean(jnp.sum((Y_target - y_pred) ** 2, axis=1))
+            )
+        else:
+            y_pred = jax.vmap(model)(X_input)
+            train_loss = float(jpc.cross_entropy_loss(y_pred, Y_target))
+        losses.append(train_loss)
+
+        model, opt_state, _, grads = make_step(
+            model, optim, opt_state, X_input, Y_target
+        )
+
+        if bp_grads is not None:
+            flat_grads = flatten_grads(grads)
+            # Convert JAX array to numpy immediately to free memory
+            bp_grads.append(np.array(flat_grads))
+            del flat_grads, grads
+
+    np.save(f"{save_dir}/losses.npy", losses)
+
+    return bp_grads
